start testing 28/41

+++ src/main/java/com/game/controller/PlayerController.java
@@ -3,40 +3,35 @@ package com.game.controller;
 import com.game.entity.Player;
 import com.game.repository.PlayerRepository;
 import com.game.service.PlayerService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
-import javax.persistence.criteria.*;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Predicate;
+import javax.persistence.criteria.Root;
 import java.util.*;
 
 @RestController
 
 public class PlayerController {
-    private static final Logger logger = LoggerFactory.getLogger(PlayerController.class);
-
     @PersistenceContext
     private EntityManager em;
     @Autowired
     PlayerService playerService;
-    @Autowired
-    private PlayerRepository playerRepository;
 
     public PlayerController(PlayerService playerService) {
         this.playerService = playerService;
     }
 
-    @GetMapping(value ="/rest/players")
+    @GetMapping(value = "/rest/players")
     public ResponseEntity<List<Player>> readAll(
             @RequestParam(required = false) String name,
             @RequestParam(required = false) String title,
@@ -52,7 +47,6 @@ public class PlayerController {
             @RequestParam(required = false) Integer maxLevel,
             @RequestParam(defaultValue = "0") int pageNumber,
             @RequestParam(defaultValue = "3") int pageSize) {
-
         order = order.toLowerCase(Locale.ROOT);
 
         CriteriaBuilder criteriaBuilder = em.getCriteriaBuilder();
@@ -62,11 +56,11 @@ public class PlayerController {
 
         Predicate criteria = criteriaBuilder.conjunction();
         if (name != null) {
-            Predicate p = criteriaBuilder.like(from.get("name"),"%"+name+"%");
+            Predicate p = criteriaBuilder.like(from.get("name"), "%" + name + "%");
             criteria = criteriaBuilder.and(criteria, p);
         }
         if (title != null) {
-            Predicate p = criteriaBuilder.like(from.get("title"), "%"+title+"%");
+            Predicate p = criteriaBuilder.like(from.get("title"), "%" + title + "%");
             criteria = criteriaBuilder.and(criteria, p);
         }
         if (race != null) {
@@ -78,115 +72,62 @@ public class PlayerController {
             criteria = criteriaBuilder.and(criteria, p);
         }
         if (banned != null) {
-            if(banned.equals("false")) {
-                Predicate p = criteriaBuilder.equal(from.get("banned"), false);
-                criteria = criteriaBuilder.and(criteria, p);
-            }else {
-                Predicate p = criteriaBuilder.equal(from.get("banned"), true);
-                criteria = criteriaBuilder.and(criteria, p);
+            Predicate p;
+            if (banned.equals("false")) {
+                p = criteriaBuilder.equal(from.get("banned"), false);
+            } else {
+                p = criteriaBuilder.equal(from.get("banned"), true);
             }
+            criteria = criteriaBuilder.and(criteria, p);
         }
-        if (minExperience != null){
+        if (minExperience != null) {
             Predicate p = criteriaBuilder.greaterThanOrEqualTo(from.get("experience"), minExperience);
             criteria = criteriaBuilder.and(criteria, p);
         }
-        if (maxExperience != null){
+        if (maxExperience != null) {
             Predicate p = criteriaBuilder.lessThanOrEqualTo(from.get("experience"), maxExperience);
             criteria = criteriaBuilder.and(criteria, p);
         }
-        if (minLevel != null){
+        if (minLevel != null) {
             Predicate p = criteriaBuilder.greaterThanOrEqualTo(from.get("level"), minLevel);
             criteria = criteriaBuilder.and(criteria, p);
         }
-        if (maxLevel != null){
+        if (maxLevel != null) {
             Predicate p = criteriaBuilder.lessThanOrEqualTo(from.get("level"), maxLevel);
             criteria = criteriaBuilder.and(criteria, p);
         }
-        if (after != null){
+        if (after != null) {
             Predicate p = criteriaBuilder.greaterThan(from.get("birthday"), new Date(after));
             criteria = criteriaBuilder.and(criteria, p);
         }
-        if (before != null){
+        if (before != null) {
             Predicate p = criteriaBuilder.lessThan(from.get("birthday"), new Date(before));
             criteria = criteriaBuilder.and(criteria, p);
         }
         criteriaQuery.where(criteria);
-        pageNumber = pageNumber*pageSize;
+        pageNumber = pageNumber * pageSize;
         TypedQuery<Player> typedQuery = em.createQuery(select);
         typedQuery.setFirstResult(pageNumber);
         typedQuery.setMaxResults(pageSize);
         List<Player> players = typedQuery.getResultList();
-        //List<Player> players;
 
-        //Pageable paging = PageRequest.of(pageNumber, pageSize, Sort.by(order));
-
-//        CriteriaBuilder cb = em.getCriteriaBuilder();
-//        CriteriaQuery<Player> criteriaQuery = cb.createQuery(Player.class);
-//        Root<Player> from = criteriaQuery.from(Player.class);
-//        //criteriaQuery.select(from);
-        //em.createQuery(criteriaQuery).setFirstResult(pageNumber).setMaxResults(pageSize);
-//        Predicate criteria = cb.conjunction();
-//        if (name != null) {
-//            Predicate p = cb.equal(from.get("name"),name);
-//            criteria = cb.and(criteria, p);
-//        }
-//        //lastName
-//        if (title != null) {
-//            Predicate p = cb.equal(from.get("title"), title);
-//            criteria = cb.and(criteria, p);
-//        }
-
-        //criteriaQuery.where(criteria);
-//        criteriaQuery.orderBy(cb.asc(from.get(order)));
-//        em.createQuery(criteriaQuery).setFirstResult(pageNumber).setMaxResults(pageSize);
-        //List<Player> result = em.createQuery(criteriaQuery).getResultList();
-        return new ResponseEntity<>( players, HttpStatus.OK);
-
-//        Page<Player> pagePlayers;
-//
-//        if (title != null)
-//            pagePlayers = playerRepository.findByTitleContaining(title, paging);
-//        else if (name != null)
-//            pagePlayers = playerRepository.findByNameContaining(name, paging);
-//        else if (banned != null)
-//            if (banned.equals("false"))
-//                pagePlayers = playerRepository.findByBanned(false, paging);
-//            else
-//                pagePlayers = playerRepository.findByBanned(true, paging);
-//        else if (race != null)
-//            pagePlayers = playerRepository.findByRaceContaining(race, paging);
-//        else if (profession != null)
-//            pagePlayers = playerRepository.findByProfessionContaining(profession, paging);
-//        else if (maxExperience != null && minExperience != null)
-//            pagePlayers = playerRepository.findByExperienceBetween(maxExperience,minExperience, paging);
-//        else if(minExperience != null )
-//            pagePlayers = playerRepository.findByExperienceGreaterThanEqual(minExperience,paging);
-//        else if(maxExperience != null )
-//            pagePlayers = playerRepository.findByExperienceLessThanEqual(maxExperience,paging);
-//        else
-//            pagePlayers =  playerRepository.findAll(paging);
-//
-//
-//
-//
-//        players = pagePlayers.getContent();
-//
-//        logger.debug("readALl()) have read "+players);
-//        return players != null && !players.isEmpty()
-//                ? new ResponseEntity<>(players, HttpStatus.OK)
-//                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(players, HttpStatus.OK);
     }
 
-    @GetMapping(value ="/rest/players/count")
-    public Integer count(
+    @GetMapping(value = "/rest/players/count")
+    public ResponseEntity<Integer> count(
             @RequestParam(required = false) String name,
             @RequestParam(required = false) String title,
             @RequestParam(required = false) String race,
             @RequestParam(required = false) String profession,
             @RequestParam(defaultValue = "id") String order,
             @RequestParam(required = false) String banned,
+            @RequestParam(required = false) Long after,
+            @RequestParam(required = false) Long before,
             @RequestParam(required = false) Integer minExperience,
             @RequestParam(required = false) Integer maxExperience,
+            @RequestParam(required = false) Integer minLevel,
+            @RequestParam(required = false) Integer maxLevel,
             @RequestParam(defaultValue = "0") int pageNumber,
             @RequestParam(defaultValue = "3") int pageSize) {
         order = order.toLowerCase(Locale.ROOT);
@@ -197,48 +138,68 @@ public class PlayerController {
         CriteriaQuery<Player> select = criteriaQuery.select(from).orderBy(criteriaBuilder.asc(from.get(order)));
         Predicate criteria = criteriaBuilder.conjunction();
         if (name != null) {
-            Predicate p = criteriaBuilder.like(from.get("name"),"%"+name+"%");
+            Predicate p = criteriaBuilder.like(from.get("name"), "%" + name + "%");
             criteria = criteriaBuilder.and(criteria, p);
         }
         if (title != null) {
-            Predicate p = criteriaBuilder.like(from.get("title"), "%"+title+"%");
+            Predicate p = criteriaBuilder.like(from.get("title"), "%" + title + "%");
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (race != null) {
+            Predicate p = criteriaBuilder.equal(from.get("race"), race);
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (profession != null) {
+            Predicate p = criteriaBuilder.equal(from.get("profession"), profession);
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (banned != null) {
+            Predicate p;
+            if (banned.equals("false")) {
+                p = criteriaBuilder.equal(from.get("banned"), false);
+            } else {
+                p = criteriaBuilder.equal(from.get("banned"), true);
+            }
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (minExperience != null) {
+            Predicate p = criteriaBuilder.greaterThanOrEqualTo(from.get("experience"), minExperience);
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (maxExperience != null) {
+            Predicate p = criteriaBuilder.lessThanOrEqualTo(from.get("experience"), maxExperience);
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (minLevel != null) {
+            Predicate p = criteriaBuilder.greaterThanOrEqualTo(from.get("level"), minLevel);
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (maxLevel != null) {
+            Predicate p = criteriaBuilder.lessThanOrEqualTo(from.get("level"), maxLevel);
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (after != null) {
+            Predicate p = criteriaBuilder.greaterThan(from.get("birthday"), new Date(after));
+            criteria = criteriaBuilder.and(criteria, p);
+        }
+        if (before != null) {
+            Predicate p = criteriaBuilder.lessThan(from.get("birthday"), new Date(before));
             criteria = criteriaBuilder.and(criteria, p);
         }
         criteriaQuery.where(criteria);
         TypedQuery<Player> typedQuery = em.createQuery(select);
         List<Player> players = typedQuery.getResultList();
-        return players.size();
-//        order = order.toLowerCase(Locale.ROOT);
-//        Pageable paging = PageRequest.of(pageNumber, pageSize, Sort.by(order));
-//
-//
-//        if (title != null)
-//            return (int)playerRepository.findByTitleContaining(title, paging).getTotalElements();
-//        else if (name != null)
-//            return (int)playerRepository.findByNameContaining(name, paging).getTotalElements();
-//        else if (banned != null)
-//            if (banned.equals("false"))
-//                return (int) playerRepository.findByBanned(false, paging).getTotalElements();
-//            else
-//                return (int) playerRepository.findByBanned(true, paging).getTotalElements();
-//        else if (race != null)
-//            return (int) playerRepository.findByRaceContaining(race, paging).getTotalElements();
-//        else if (profession != null)
-//            return (int) playerRepository.findByProfessionContaining(profession, paging).getTotalElements();
-//        else if (maxExperience != null && minExperience != null)
-//            return (int) playerRepository.findByExperienceBetween(maxExperience,minExperience, paging).getTotalElements();
-//        else if (maxExperience != null )
-//            return (int) playerRepository.findByExperienceLessThanEqual(maxExperience, paging).getTotalElements();
-//        else if (minExperience != null )
-//            return (int) playerRepository.findByExperienceGreaterThanEqual(minExperience, paging).getTotalElements();
-//        else return (int) playerRepository.findAll(paging).getTotalElements();
+        int res = players.size();
+        return new ResponseEntity<>(res, HttpStatus.OK);
     }
 
     @GetMapping("/rest/players/{id}")
     public ResponseEntity<Player> get(@PathVariable Long id) {
         try {
+            if(id < 0){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             Player player = playerService.getPlayer(id);
-            logger.debug("get() got player "+player);
             return new ResponseEntity<>(player, HttpStatus.OK);
         } catch (NoSuchElementException e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -246,34 +207,47 @@ public class PlayerController {
     }
 
     @PostMapping("/rest/players")
-    public ResponseEntity<?> create (@RequestBody Player player) {
-        player.setLevel((int) ((Math.sqrt(2500
-                        +200*player.getExperience())-50)/100));
-        player.setUntilNextLevel(50*(player.getLevel()+1)*(player.getLevel()+2)- player.getExperience());
+    public ResponseEntity<?> create(@RequestBody Player player) {
+        Date twoThousand = new GregorianCalendar(2000,Calendar.JANUARY,1).getTime();
+        Date threeThousand = new GregorianCalendar(2999,Calendar.DECEMBER,31).getTime();
+            if(player.getName().isEmpty() || player.getName().length() > 12 || player.getTitle().isEmpty() || player.getExperience() == null
+                    || player.getTitle().length() > 30 || player.getRace().isEmpty() || player.getProfession().isEmpty() ||
+                    player.getExperience() < 0 || player.getExperience() > 10000
+                    || player.getBirthday().before(twoThousand)    || player.getBirthday().after(threeThousand)){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+
+        player.setLevel((int) ((Math.sqrt(2500 + 200 * player.getExperience()) - 50) / 100));
+        player.setUntilNextLevel(50 * (player.getLevel() + 1) * (player.getLevel() + 2) - player.getExperience());
         playerService.create(player);
-        logger.debug("create() created player "+ player);
         return new ResponseEntity<>(player, HttpStatus.OK);
     }
 
     @PostMapping("/rest/players/{id}")
     public ResponseEntity<?> update(@RequestBody Player player, @PathVariable Long id) {
         try {
-            logger.debug("update() updating "+ player);
+            if(id < 0){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
             Player existPlayer = playerService.getPlayer(id);
-           player.setId(id);
+            player.setId(id);
 
             return create(player);
         } catch (NoSuchElementException e) {
-            logger.error("update() failed" + e);
-            return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
     @DeleteMapping("/rest/players/{id}")
-    public void delete(@PathVariable Long id) {
-        logger.debug("deleting player" + id);
-        playerService.deletePlayer(id);
-        logger.debug("player" + id+"deleted");
+    public ResponseEntity<?> delete(@PathVariable Long id) {
+        try {
+            if(id < 0){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            playerService.deletePlayer(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }catch (NoSuchElementException e){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
-
 }
